Revert "Add Stackdriver Log Viewer Tool Window."
This reverts commit 1c270596a49d39181e142b93ad74db94106b5df2.
